Move files into a more appropriate folder
